Starting to rewrite in Scala with Lagom's recommended structure.
Key changes is that in Scala, Guice isn't used as a dependency injection
system, macwire is. In this context, the implementation of services need
to extend LagomApplication and attach multiple "components" according to
what is needed. In the case of Users, CassandraPersistenceComponents is
used to store User data separately from the event journal, and it's seen
that the UserRepository is wired before the UserServiceImpl is wired.

There's a fair amount of work remaining to keep migrating the remainder
of the implementation of these microservices, but it is going to be done
eventually.

Signed-off-by: Gabriel Harris-Rouquette <gabizou@me.com>